requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
youtube_dl>=2021.12.17
pdf2image>=1.16.3
openai>=1.3.0
